# 开发技术文档/02-API接口定义文档/openapi.yaml
openapi: "3.0.3"
info:
  title: 知微 ziwi SaaS 平台 API
  description: |
    知微(SaaS)平台 — Phase 1 核心业务 API 文档

    ## 认证方式
    - **Bearer JWT**: `Authorization: Bearer <token>` — 用户身份认证
    - **X-Tenant-Id**: 请求头 — 多租户识别（SaaS 模式必传）
    - **X-Deploy-Mode**: 请求头 — 部署模式: saas / onprem / multi-factory
    - **X-Factory-Id**: 请求头 — 工厂ID（multi-factory 模式必传）
    - **X-License-Key**: 请求头 — 许可证密钥（onprem 模式必传）

    ## API 版本
    - URL 路径版本: `/api/v1/{resource}`
    - 同一接口保留 2 个版本，v1 废弃前 3 个月发送 `X-API-Deprecated` 头

    ## 分页规范
    - `page`: 从 1 开始
    - `page_size`: 默认 20，最大 100
    - `total`: 总记录数
    - 响应格式: `{ "items": [], "total": 0, "page": 1, "page_size": 20 }`

    ## 错误响应格式
    ```json
    {
      "code": "400-1001",
      "message": "参数校验失败：必填字段缺失",
      "request_id": "req_abc123"
    }
    ```

    ## 错误码清单
    | code | 说明 |
    |------|------|
    | 400-0000 | 请求参数错误 |
    | 400-1001 | 参数校验失败：必填字段缺失 |
    | 400-1002 | 参数校验失败：格式错误 |
    | 400-1003 | 参数校验失败：值超出允许范围 |
    | 400-2001 | 业务规则冲突：库存不足 |
    | 400-2002 | 业务规则冲突：状态不允许此操作 |
    | 400-2003 | 业务规则冲突：数据已存在（重复） |
    | 401-0000 | 未认证 |
    | 401-1001 | Token 无效或已过期 |
    | 401-1002 | Token 格式错误 |
    | 401-1003 | OAuth 授权失败 |
    | 403-0000 | 无权限 |
    | 403-1001 | 模块未购买（租户未启用该模块） |
    | 403-1002 | 用户角色无此操作权限 |
    | 403-1003 | 数据范围权限不足（不可跨工厂操作） |
    | 404-0000 | 资源不存在 |
    | 404-1001 | 查询的资源 ID 不存在 |
    | 404-1002 | 关联的父资源不存在 |
    | 409-0000 | 数据冲突 |
    | 409-1001 | 并发修改冲突（乐观锁版本号不一致） |
...
    | 429-0000 | 请求过频繁 |
...
servers:
...
paths:
...
  /api/v1/roles:
    get:
      tags: ["M00-角色管理"]
      summary: 角色列表
      operationId: listRoles
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: 角色列表
    post:
      tags: ["M00-角色管理"]
      summary: 创建角色
      operationId: createRole
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/roles/{role_id}:
    get:
      tags: ["M00-角色管理"]
      summary: 角色详情
      operationId: getRole
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M00-角色管理"]
      summary: 更新角色
      operationId: updateRole
      security:
        - BearerAuth: []
        - TenantId: []
    delete:
      tags: ["M00-角色管理"]
      summary: 删除角色
      operationId: deleteRole
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/roles/{role_id}/permissions:
    get:
      tags: ["M00-角色管理"]
      summary: 查询角色权限
      operationId: getRolePermissions
      security:
        - BearerAuth: []
        - TenantId: []
    put:
      tags: ["M00-角色管理"]
      summary: 分配角色权限
      operationId: assignRolePermissions
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/roles/{role_id}/users:
    get:
      tags: ["M00-角色管理"]
      summary: 查询角色关联用户
      operationId: getRoleUsers
      security:
        - BearerAuth: []
        - TenantId: []
    post:
      tags: ["M00-角色管理"]
      summary: 为角色添加用户
      operationId: addRoleUser
      security:
        - BearerAuth: []
        - TenantId: []

  # ==========================================
  # M00 — 权限
  # ==========================================
...
  /api/v1/inspection-orders:
    get:
      tags: ["M03-检验"]
      summary: 检验单列表
      operationId: listInspectionOrders
      security:
        - BearerAuth: []
        - TenantId: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: inspection_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: result
          in: query
          schema:
            type: string
        - name: batch_no
          in: query
          schema:
            type: string
        - name: factory_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 检验单列表
    post:
      tags: ["M03-检验"]
      summary: 创建检验单
      operationId: createInspectionOrder
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/inspection-orders/{order_id}:
    get:
      tags: ["M03-检验"]
      summary: 检验单详情(含结果)
      operationId: getInspectionOrder
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/inspection-orders/{order_id}/results:
    post:
      tags: ["M03-检验"]
      summary: 提交检验结果
      operationId: submitInspectionResults
      security:
        - BearerAuth: []
        - TenantId: []

  # ==========================================
  # M03 — 合格证
  # ==========================================
...
  /api/v1/quality-reports:
    get:
      tags: ["M03-质量报告"]
      summary: 质量报告列表
      operationId: listQualityReports
      security:
        - BearerAuth: []
        - TenantId: []
    post:
      tags: ["M03-质量报告"]
      summary: 生成质量报告
      operationId: generateQualityReport
      security:
        - BearerAuth: []
        - TenantId: []
  /api/v1/quality-reports/{report_id}:
    get:
      tags: ["M03-质量报告"]
      summary: 质量报告详情
      operationId: getQualityReport
      security:
        - BearerAuth: []
        - TenantId: []

  # ==========================================
  # M05 — 安灯
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer Token，从登录接口获取"
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: "租户ID（SaaS模式必传）"
    DeployMode:
      type: apiKey
      in: header
      name: X-Deploy-Mode
      description: "部署模式: saas / onprem / multi-factory"
    FactoryId:
      type: apiKey
      in: header
      name: X-Factory-Id
      description: "工厂ID（multi-factory模式必传）"
    LicenseKey:
      type: apiKey
      in: header
      name: X-License-Key
      description: "许可证密钥（onprem模式必传）"
...
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码，从 1 开始
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 每页条数，默认 20，最大 100
      schema:
        type: integer
        default: 20
        maximum: 100
        minimum: 1
...